feat: new page objectif can read data now
